update paths for moved components
